[Write API]: Code refactor. Added Service class.
