+++ project/src/main/kotlin/org/ionproject/core/writeApi/common/Uri.kt
@@ -1,6 +1,6 @@
 package org.ionproject.core.writeApi.common
 
 object Uri {
-  const val version = "/v0"
-  const val insertClassSectionEvents = "${version}/insertClassSectionEvents"
+    const val version = "/v0"
+    const val insertClassSectionEvents = "${version}/insertClassSectionEvents"
 }
