+++ project/src/main/kotlin/org/ionproject/core/writeApi/insertClassSectionEvents/sql/InsertClassSectionEventsData.kt
@@ -1,31 +1,31 @@
 package org.ionproject.core.writeApi.insertClassSectionEvents.sql
 
 internal object InsertClassSectionEventsData {
-  const val SCHEMA = "dbo"
+    const val SCHEMA = "dbo"
 
-  const val UPSERT_SCHOOL_PROC = "sp_createOrReplaceSchool"
-  const val UPSERT_COURSE_PROC = "sp_createOrReplaceCourse"
-  const val CREATE_CLASS_SECTION_EVENT_PROC = "sp_createClassSectionEvent"
+    const val UPSERT_SCHOOL_PROC = "sp_createOrReplaceSchool"
+    const val UPSERT_COURSE_PROC = "sp_createOrReplaceCourse"
+    const val CREATE_CLASS_SECTION_EVENT_PROC = "sp_createClassSectionEvent"
 
-  const val SCHOOL_NAME_PARAM = "schoolName"
-  const val SCHOOL_ACR_PARAM = "schoolAcr"
-  const val COURSE_NAME_PARAM = "courseName"
-  const val COURSE_ACR_PARAM = "courseAcr"
-  const val PROGRAMME_NAME_PARAM = "programmeName"
-  const val PROGRAMME_ACR_PARAM = "programmeAcr"
-  const val PROGRAMME_TERM_SIZE_PARAM = "programmeTermSize"
-  const val CALENDAR_TERM_PARAM = "calTerm"
-  const val CALENDAR_SECTION_PARAM = "calendarSection"
-  const val LANGUAGE_PARAM = "lang"
-  const val CATEGORY_PARAM = "categ"
-  const val SUMMARY_PARAM = "summary"
-  const val DESCRIPTION_PARAM = "description"
-  const val DTSTART_PARAM = "dtstart"
-  const val DTEND_PARAM = "dtend"
-  const val WEEK_DAYS_PARAM = "week_days"
-  const val LOCATION_PARAM = "location"
+    const val SCHOOL_NAME_PARAM = "schoolName"
+    const val SCHOOL_ACR_PARAM = "schoolAcr"
+    const val COURSE_NAME_PARAM = "courseName"
+    const val COURSE_ACR_PARAM = "courseAcr"
+    const val PROGRAMME_NAME_PARAM = "programmeName"
+    const val PROGRAMME_ACR_PARAM = "programmeAcr"
+    const val PROGRAMME_TERM_SIZE_PARAM = "programmeTermSize"
+    const val CALENDAR_TERM_PARAM = "calTerm"
+    const val CALENDAR_SECTION_PARAM = "calendarSection"
+    const val LANGUAGE_PARAM = "lang"
+    const val CATEGORY_PARAM = "categ"
+    const val SUMMARY_PARAM = "summary"
+    const val DESCRIPTION_PARAM = "description"
+    const val DTSTART_PARAM = "dtstart"
+    const val DTEND_PARAM = "dtend"
+    const val WEEK_DAYS_PARAM = "week_days"
+    const val LOCATION_PARAM = "location"
 
-  const val CALL_UPSERT_SCHOOL = """
+    const val CALL_UPSERT_SCHOOL = """
     CALL $SCHEMA.$UPSERT_SCHOOL_PROC(
         :$SCHOOL_NAME_PARAM,
         :$SCHOOL_ACR_PARAM,
@@ -35,7 +35,7 @@ internal object InsertClassSectionEventsData {
         :$CALENDAR_TERM_PARAM)
   """
 
-  const val CALL_UPSERT_COURSE = """
+    const val CALL_UPSERT_COURSE = """
     CALL $SCHEMA.$UPSERT_COURSE_PROC(
         :$COURSE_NAME_PARAM,
         :$COURSE_ACR_PARAM,
@@ -45,7 +45,7 @@ internal object InsertClassSectionEventsData {
         :$CATEGORY_PARAM)
   """
 
-  const val CALL_CREATE_EVENT = """
+    const val CALL_CREATE_EVENT = """
     CALL $SCHEMA.$CREATE_CLASS_SECTION_EVENT_PROC(
         :$COURSE_NAME_PARAM,
         :$COURSE_ACR_PARAM,
